+++ AsocialCartography/Plugin.cs
@@ -1,9 +1,10 @@
-﻿using System.IO;
-using System.Reflection;
-using BepInEx;
+﻿using BepInEx;
 using BepInEx.Configuration;
 using BepInEx.Logging;
 using HarmonyLib;
+using System;
+using System.IO;
+using System.Reflection;
 
 namespace VentureValheim.AsocialCartography;
 
@@ -11,7 +12,7 @@ namespace VentureValheim.AsocialCartography;
 public class AsocialCartographyPlugin : BaseUnityPlugin
 {
     private const string ModName = "AsocialCartography";
-    private const string ModVersion = "0.3.0";
+    private const string ModVersion = "0.3.1";
     private const string Author = "com.orianaventure.mod";
     private const string ModGUID = Author + "." + ModName;
     private static string ConfigFileName = ModGUID + ".cfg";
@@ -27,12 +28,14 @@ public class AsocialCartographyPlugin : BaseUnityPlugin
     private static ConfigEntry<bool> CE_ReceivePins = null!;
     private static ConfigEntry<bool> CE_IgnoreBossPins = null!;
     private static ConfigEntry<bool> CE_IgnoreHildirPins = null!;
+    private static ConfigEntry<string> CE_IgnoredCustomPins = null!;
     private static ConfigEntry<float> CE_ReceivePinRadius = null!;
 
     public static bool GetAddPins() => CE_AddPins.Value;
     public static bool GetIgnoreBossPins() => CE_IgnoreBossPins.Value;
     public static bool GetIgnoreHildirPins() => CE_IgnoreHildirPins.Value;
     public static bool GetReceivePins() => CE_ReceivePins.Value;
+    public static string GetIgnoredCustomPins() => CE_IgnoredCustomPins.Value;
     public static float GetReceivePinRadius() => CE_ReceivePinRadius.Value;
 
     private void AddConfig<T>(string key, string section, string description, bool synced, T value, ref ConfigEntry<T> configEntry)
@@ -62,6 +65,8 @@ public class AsocialCartographyPlugin : BaseUnityPlugin
             false, true, ref CE_IgnoreBossPins);
         AddConfig("IgnoreHildirPins", general, "False to include hildir map pins in the above configs (boolean).",
             false, true, ref CE_IgnoreHildirPins);
+        AddConfig("IgnoredCustomPins", general, "List of map pins by integer id to include in the above configs (comma-separated list of integers).",
+            false, "", ref CE_IgnoredCustomPins);
         AddConfig("ReceivePinRadius", general, "Overlap radius that must be exceeded to receive a pin from the map table. Default in vanilla is 1 (float).",
             false, 50f, ref CE_ReceivePinRadius);
 
@@ -70,6 +75,7 @@ public class AsocialCartographyPlugin : BaseUnityPlugin
         Assembly assembly = Assembly.GetExecutingAssembly();
         HarmonyInstance.PatchAll(assembly);
         SetupWatcher();
+        AsocialCartography.UpdateConfigurations();
     }
 
     private void OnDestroy()
@@ -79,7 +85,9 @@ public class AsocialCartographyPlugin : BaseUnityPlugin
 
     private void SetupWatcher()
     {
+        _lastReloadTime = DateTime.Now;
         FileSystemWatcher watcher = new(BepInEx.Paths.ConfigPath, ConfigFileName);
+        // Due to limitations of technology this can trigger twice in a row
         watcher.Changed += ReadConfigValues;
         watcher.Created += ReadConfigValues;
         watcher.Renamed += ReadConfigValues;
@@ -88,17 +96,28 @@ public class AsocialCartographyPlugin : BaseUnityPlugin
         watcher.EnableRaisingEvents = true;
     }
 
+    private DateTime _lastReloadTime;
+    private const long RELOAD_DELAY = 10000000; // One second
+
     private void ReadConfigValues(object sender, FileSystemEventArgs e)
     {
-        if (!File.Exists(ConfigFileFullPath)) return;
+        var now = DateTime.Now;
+        var time = now.Ticks - _lastReloadTime.Ticks;
+        if (!File.Exists(ConfigFileFullPath) || time < RELOAD_DELAY) return;
+
         try
         {
-            AsocialCartographyLogger.LogDebug("Attempting to reload configuration...");
+            AsocialCartographyLogger.LogInfo("Attempting to reload configuration...");
             Config.Reload();
         }
         catch
         {
             AsocialCartographyLogger.LogError($"There was an issue loading {ConfigFileName}");
+            return;
         }
+
+        _lastReloadTime = now;
+
+        AsocialCartography.UpdateConfigurations();
     }
 }